+++ src/logic/model/SThread.java
@@ -10,14 +10,9 @@ import except.LoginCredentialException;
 import except.LoginExistsException;
 import except.NotRegisteredException;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-import java.io.OutputStream;
 import java.net.Socket;
-import java.sql.SQLException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import logic.ControllerThread;
 import logic.DBReaderFactory;
 import logic.objects.User;
